Correct locking and make new a const fn
This rewrites the locking, moving the futures::Mutex to a dedicated Box
that is only present during initialization.

It also fixes unsafety in get(): it was previously possible for the
Option to be set to Some() during set() before set() completes,
resulting in a reference to an incompletely written copy of T being
returned.